keepalive: windows fix
